refactor: Move has_web_form permission code to permissions.py
