+++ frappe/permissions.py
@@ -549,3 +549,26 @@ def filter_allowed_docs_for_doctype(user_permissions, doctype, with_default_doc=
 def push_perm_check_log(log):
 	if frappe.flags.get('has_permission_check_logs') == None: return
 	frappe.flags.get('has_permission_check_logs').append(_(log))
+
+def has_web_form_permission(doctype, name, ptype='read'):
+	if frappe.session.user=="Guest":
+		return False
+
+	# owner matches
+	elif frappe.db.get_value(doctype, name, "owner")==frappe.session.user:
+		return True
+
+	elif frappe.has_website_permission(name, ptype=ptype, doctype=doctype):
+		return True
+
+	elif check_webform_perm(doctype, name):
+		return True
+
+	else:
+		return False
+
+def check_webform_perm(doctype, name):
+	doc = frappe.get_doc(doctype, name)
+	if hasattr(doc, "has_webform_permission"):
+		if doc.has_webform_permission():
+			return True
